+++ tests/TestLongOperand.java
@@ -6,7 +6,7 @@ public class TestLongOperand {
         // cache pool bounds: [-128; 127]
         for (long x = (-128 - 1); x <= (127 + 1); x++) {
             LongOperand wrapper_x = LongOperand.valueOf(x);
-            if (wrapper_x.intValue() != x) {
+            if (wrapper_x.longValue() != x) {
                 throw new AssertionError(x);
             }
         }
